requirements.txt
# Web Scraping
selenium==4.16.0
webdriver-manager==4.0.1
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==5.1.0
requests==2.31.0

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL
# pymongo==4.6.1  # Caso prefira MongoDB

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
pandas==2.1.4
schedule==1.2.0

# Monitoring & Logs
loguru==0.7.2
